+++ code/test/sixsq/nuvla/server/resources/nuvlabox_status_1_lifecycle_test.clj
@@ -156,6 +156,15 @@
              (ltu/body->edn)
              (ltu/is-status 403))
 
+         ;; admin edition doesn't set online flag
+         (-> session-admin
+             (request state-url
+                      :request-method :put
+                      :body (json/write-str {}))
+             (ltu/body->edn)
+             (ltu/is-status 200)
+             (ltu/is-key-value :online nil))
+
          ;; nuvlabox user is able to update nuvlabox-status
          (-> session-nb
              (request state-url
@@ -163,7 +172,17 @@
                       :body (json/write-str {:resources resources-updated}))
              (ltu/body->edn)
              (ltu/is-status 200)
-             (ltu/is-key-value :resources resources-updated))
+             (ltu/is-key-value :resources resources-updated)
+             (ltu/is-key-value :online true))
+
+         ;; admin edition can set online flag
+         (-> session-admin
+             (request state-url
+                      :request-method :put
+                      :body (json/write-str {:online false}))
+             (ltu/body->edn)
+             (ltu/is-status 200)
+             (ltu/is-key-value :online false))
 
          ;; verify that the update was written to disk
          (-> session-nb